+++ src/main/java/com/team/comma/spotify/track/service/TrackService.java
@@ -1,0 +1,22 @@
+package com.team.comma.spotify.track.service;
+
+import com.team.comma.spotify.playlist.dto.PlaylistTrackArtistResponse;
+import com.team.comma.spotify.track.domain.TrackArtist;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class TrackService {
+
+    public List<PlaylistTrackArtistResponse> createArtistResponse(List<TrackArtist> artistList){
+        List<PlaylistTrackArtistResponse> result = new ArrayList<>();
+        for (TrackArtist artist : artistList){
+            result.add(PlaylistTrackArtistResponse.of(artist));
+        }
+        return result;
+    }
+}
